Address PR feedback - move to crypto, remove virtual, and misc.
