+++ sdk/core/azure-core/CHANGELOG.md
@@ -5,7 +5,7 @@
 ### New Features
 
 - Added support for HTTP conditional requests `MatchConditions` and `RequestConditions`.
-- Added the `Hash` base class and MD5 hashing APIs to the `Azure::Core` namespace available from `azure/core/hash.hpp`.
+- Added the `Hash` base class and MD5 hashing APIs to the `Azure::Core::Cryptography` namespace available from `azure/core/cryptography/hash.hpp`.
 
 ### Breaking Changes
