+++ sdk/core/azure-core/test/ut/simplified_header.cpp
@@ -23,7 +23,7 @@ TEST(SimplifiedHeader, core)
   EXPECT_NO_THROW(Azure::Core::DateTime(2020, 11, 03, 15, 30, 44));
   EXPECT_NO_THROW(Azure::Core::ETag e);
   EXPECT_NO_THROW(Azure::Core::Base64Decode("foo"));
-  EXPECT_NO_THROW(Azure::Core::Md5Hash m);
+  EXPECT_NO_THROW(Azure::Core::Cryptography::Md5Hash m);
   EXPECT_NO_THROW(Azure::Core::Http::RawResponse r(
       1, 1, Azure::Core::Http::HttpStatusCode::Accepted, "phrase"));
   EXPECT_NO_THROW(Azure::Core::MatchConditions mc);
